big update
BUGFIXES
Recurring invoices are now displaying in the proper sort order.
The Tax Summary Report has been adjusted so it produces more sensible
results.

ENHANCEMENTS
The invoice and quote edit screen has been optimized for better display
and interaction.
Invoices and quotes can now be emailed to multiple addresses separated
by a comma.
Invoices and quotes in foreign currencies are now displayed in the
foreign currency and not the base currency.
Default invoice and quote templates can now be set per client and / or
quote/invoice.
Uploaded logos will now display as the size of the actual image uploaded
without needing to specify the size in the template.
Added new favicon.