big update. read description
Enhancements

- Contrary to the fact that the primary purpose of HubPay is to generate
invoices to send to your clients, previous versions have shipped with a
rather ugly default template. The default templates have been redesigned
to provide a much cleaner and professional look for your invoices and
give you a much better base for customization.
- In previous releases, taxes always displayed sort of weird in the
summary area of a quote or invoice. Taxes are now summarized properly in
the totals area of the default quote and invoice templates. Those of you
using customized templates should review the new default template for
the changes to the tax area.
- Payment receipts can now be sent manually from the Payments screen.
- A new Client Statement report provides a report including all open
invoices for the specified client. The other existing reports have
received minor updates as well.
- The Client detail screen now has a Payments tab.
- Performed a lot of code cleanup and removal of unnecessary files and
resources.
- Quotes and invoices are now sorted in a more natural manner.

Bugfixes
- A default temp directory is now specified for PDF generation to
alleviate permission issues a number of environments had when attempting
to write to the default system temp directory.